Move category field picker out of deprecated widgets directory
