+++ frontend/src/metabase/core/components/FormCategoryInput/CategoryFieldPicker.tsx
